Removed Version 1_0 from AD Extension
