+++ src/main/java/org/opensearch/ad/AnomalyDetectorExtension.java
@@ -103,8 +103,7 @@ public class AnomalyDetectorExtension implements Extension {
 
     public OpenSearchClient getClient() {
         SDKClient sdkClient = new SDKClient();
-        OpenSearchClient client = sdkClient
-            .initializeClient(settings.getOpensearchAddress(), Integer.parseInt(settings.getOpensearchPort()));
+        OpenSearchClient client = sdkClient.initializeClient("localhost", 9200);
         return client;
     }
