+++ src/main/java/org/opensearch/ad/cluster/ADVersionUtil.java
@@ -43,7 +43,4 @@ public class ADVersionUtil {
         return normalizedVersion.toString();
     }
 
-    public static boolean compatibleWithVersionOnOrAfter1_1(Version adVersion) {
-        return adVersion != null && adVersion.onOrAfter(Version.V_1_1_0);
-    }
 }
