+++ src/main/java/org/opensearch/ad/model/ADTaskProfile.java
@@ -20,7 +20,6 @@ import java.util.Objects;
 
 import org.opensearch.Version;
 import org.opensearch.ad.annotation.Generated;
-import org.opensearch.ad.cluster.ADVersionUtil;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.io.stream.Writeable;
@@ -181,7 +180,7 @@ public class ADTaskProfile implements ToXContentObject, Writeable {
         out.writeOptionalInt(thresholdModelTrainingDataSize);
         out.writeOptionalLong(modelSizeInBytes);
         out.writeOptionalString(nodeId);
-        if (ADVersionUtil.compatibleWithVersionOnOrAfter1_1(adVersion)) {
+        if (adVersion != null) {
             out.writeOptionalString(taskId);
             out.writeOptionalString(adTaskType);
             out.writeOptionalInt(detectorTaskSlots);
