+++ src/main/java/org/opensearch/ad/rest/RestCreateDetectorAction.java
@@ -94,6 +94,7 @@ public class RestCreateDetectorAction implements ExtensionRestHandler {
 
         IndexRequest<AnomalyDetector> indexRequest = new IndexRequest.Builder<AnomalyDetector>()
             .index(ANOMALY_DETECTORS_INDEX)
+            .id("1")
             .document(detector)
             .build();
         IndexResponse indexResponse = sdkClient.index(indexRequest);
