+++ src/main/java/org/opensearch/ad/transport/EntityProfileRequest.java
@@ -23,7 +23,6 @@ import org.opensearch.ad.constant.CommonErrorMessages;
 import org.opensearch.ad.constant.CommonName;
 import org.opensearch.ad.model.Entity;
 import org.opensearch.ad.model.EntityProfileName;
-import org.opensearch.ad.util.Bwc;
 import org.opensearch.common.Strings;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
@@ -41,17 +40,7 @@ public class EntityProfileRequest extends ActionRequest implements ToXContentObj
     public EntityProfileRequest(StreamInput in) throws IOException {
         super(in);
         adID = in.readString();
-        if (Bwc.supportMultiCategoryFields(in.getVersion())) {
-            entityValue = new Entity(in);
-        } else {
-            // entity profile involving an old node won't work. Read
-            // EntityProfileTransportAction.doExecute for details. Read
-            // a string to not cause EOF exception.
-            // Cannot assign null to entityValue as old node has no logic to
-            // deal with a null entity.
-            String oldFormatEntityString = in.readString();
-            entityValue = Entity.createSingleAttributeEntity(CommonName.EMPTY_FIELD, oldFormatEntityString);
-        }
+        entityValue = new Entity(in);
         int size = in.readVInt();
         profilesToCollect = new HashSet<EntityProfileName>();
         if (size != 0) {
@@ -84,14 +73,8 @@ public class EntityProfileRequest extends ActionRequest implements ToXContentObj
     public void writeTo(StreamOutput out) throws IOException {
         super.writeTo(out);
         out.writeString(adID);
-        if (Bwc.supportMultiCategoryFields(out.getVersion())) {
-            entityValue.writeTo(out);
-        } else {
-            // entity profile involving an old node won't work. Read
-            // EntityProfileTransportAction.doExecute for details. Write
-            // a string to not cause EOF exception.
-            out.writeString(entityValue.toString());
-        }
+        entityValue.writeTo(out);
+
         out.writeVInt(profilesToCollect.size());
         for (EntityProfileName profile : profilesToCollect) {
             out.writeEnum(profile);
