+++ src/main/java/org/opensearch/ad/transport/EntityProfileResponse.java
@@ -19,9 +19,7 @@ import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.apache.commons.lang.builder.ToStringBuilder;
 import org.opensearch.action.ActionResponse;
 import org.opensearch.ad.constant.CommonName;
-import org.opensearch.ad.model.ModelProfile;
 import org.opensearch.ad.model.ModelProfileOnNode;
-import org.opensearch.ad.util.Bwc;
 import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import org.opensearch.common.xcontent.ToXContentObject;
@@ -82,14 +80,7 @@ public class EntityProfileResponse extends ActionResponse implements ToXContentO
         lastActiveMs = in.readLong();
         totalUpdates = in.readLong();
         if (in.readBoolean()) {
-            if (Bwc.supportMultiCategoryFields(in.getVersion())) {
-                modelProfile = new ModelProfileOnNode(in);
-            } else {
-                // we don't have model information from old node
-                ModelProfile profile = new ModelProfile(in);
-                modelProfile = new ModelProfileOnNode("", profile);
-            }
-
+            modelProfile = new ModelProfileOnNode(in);
         } else {
             modelProfile = null;
         }
@@ -118,12 +109,7 @@ public class EntityProfileResponse extends ActionResponse implements ToXContentO
         out.writeLong(totalUpdates);
         if (modelProfile != null) {
             out.writeBoolean(true);
-            if (Bwc.supportMultiCategoryFields(out.getVersion())) {
-                modelProfile.writeTo(out);
-            } else {
-                ModelProfile oldFormatModelProfile = modelProfile.getModelProfile();
-                oldFormatModelProfile.writeTo(out);
-            }
+            modelProfile.writeTo(out);
         } else {
             out.writeBoolean(false);
         }
