+++ src/main/resources/ad-extension.yml
@@ -1,5 +1,3 @@
 extensionName: ad-extension
 hostAddress: 127.0.0.1
 hostPort: 4532
-opensearchAddress: 127.0.01
-opensearchPort: 9200
