+++ src/test/java/org/opensearch/ad/EntityProfileRunnerTests.java
@@ -263,6 +263,18 @@ public class EntityProfileRunnerTests extends AbstractADTest {
         assertTrue(inProgressLatch.await(100, TimeUnit.SECONDS));
     }
 
+    public void testEmptyModelProfile() throws IOException {
+        ModelProfile modelProfile = new ModelProfile(modelId, null, modelSize);
+        BytesStreamOutput output = new BytesStreamOutput();
+        modelProfile.writeTo(output);
+        StreamInput streamInput = output.bytes().streamInput();
+        ModelProfile readResponse = new ModelProfile(streamInput);
+        assertEquals("serialization has the wrong model id", modelId, readResponse.getModelId());
+        assertTrue("serialization has null entity", null == readResponse.getEntity());
+        assertEquals("serialization has the wrong model size", modelSize, readResponse.getModelSizeInBytes());
+
+    }
+
     @SuppressWarnings("unchecked")
     public void testJobIndexNotFound() throws InterruptedException {
         setUpExecuteEntityProfileAction(InittedEverResultStatus.INITTED);
