+++ src/test/java/org/opensearch/ad/transport/CronTransportActionTests.java
@@ -21,7 +21,6 @@ import java.util.function.Function;
 
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.opensearch.Version;
 import org.opensearch.action.support.ActionFilters;
 import org.opensearch.ad.AbstractADTest;
@@ -33,7 +32,6 @@ import org.opensearch.ad.feature.FeatureManager;
 import org.opensearch.ad.ml.EntityColdStarter;
 import org.opensearch.ad.ml.ModelManager;
 import org.opensearch.ad.task.ADTaskManager;
-import org.opensearch.ad.util.Bwc;
 import org.opensearch.cluster.ClusterName;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.service.ClusterService;
@@ -53,11 +51,6 @@ import com.google.gson.JsonElement;
 public class CronTransportActionTests extends AbstractADTest {
     private CronTransportAction action;
     private String localNodeID;
-
-    @BeforeClass
-    public static void setUpBeforeClass() {
-        Bwc.DISABLE_BWC = false;
-    }
 
     @Override
     @Before
@@ -100,10 +93,10 @@ public class CronTransportActionTests extends AbstractADTest {
 
         CronNodeRequest nodeRequest = new CronNodeRequest();
         BytesStreamOutput nodeRequestOut = new BytesStreamOutput();
-        nodeRequestOut.setVersion(Version.V_1_0_0);
+        nodeRequestOut.setVersion(Version.V_2_0_0);
         nodeRequest.writeTo(nodeRequestOut);
         StreamInput siNode = nodeRequestOut.bytes().streamInput();
-        siNode.setVersion(Version.V_1_0_0);
+        siNode.setVersion(Version.V_2_0_0);
 
         CronNodeRequest nodeResponseRead = new CronNodeRequest(siNode);
